Allow overwriting answers
Added suport for overwriting the answers with msg.answer. msg.answer could be an array like ["a","b","c"]. Then the value (a,b,c) will be the returned payload. It also could be
[
{"text":"a", "callback":"cA"},
{"text":"b", "callback":"cB"},
{"text":"c", "callback":"cC"}
]
Then the callback will be returned.
In that case only one out node will be displayed, exceped when  timeout is enabled. In that case two nodes will be displayed.